Create main
This code uses a random forest regressor to train a model that can predict the compatibility score between two people based on their characteristics. The model is trained on a dataset of dating data, and the mean absolute error is used to evaluate the model's performance.

Of course, this is just one example of how you could create a dating app using Python. There are many other algorithms and techniques that you could use to improve the accuracy and functionality of your app. For example, you could use a neural network to train a more sophisticated model, or you could incorporate additional features, such as user preferences and user behavior, to improve the app's performance.